+++ src/main/java/backend/model/qa/QuestionModel.java
@@ -1,5 +1,6 @@
 package backend.model.qa;
 
+import backend.model.qa.common.ForumPost;
 import backend.model.vote.QuestionVoteModel;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -12,75 +13,32 @@ import java.util.Set;
  */
 
 @Entity
-public class QuestionModel{
+public class QuestionModel extends ForumPost {
 
-    @Id
+    @OneToMany(cascade = CascadeType.ALL)
     @JsonIgnore
-    private long id;
-    @OneToMany
-    @JsonIgnore
-    private Set<AnswerModel> answers = new HashSet<>();
+    private Set<AnswerModel> answerModel = new HashSet<>();
     @OneToOne(cascade = CascadeType.ALL)
     @JsonIgnore
     private QuestionVoteModel votes;
-    @JsonIgnore
-    private String dateTime;
 
-    //TODO replace with User issue#8
-    private String author;
-    private String postMessage;
-
-    public QuestionModel(String postMessage, String author, String dateTime) {
-        this.postMessage= postMessage;
-        this.author = author;
-        this.dateTime = dateTime;
+    public QuestionModel(String message, String author, String dateTime) {
+        super(message, author, dateTime);
     }
 
-    public QuestionModel(){ //for jpa
-
+    public QuestionModel() { //for jpa
+        super("", "", "");
     }
 
-    public void setVotes(QuestionVoteModel votes){
+    public void setVotes(QuestionVoteModel votes) {
         this.votes = votes;
     }
 
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public Set<AnswerModel> getAnswers() {
-        return answers;
+    public Set<AnswerModel> getAnswerModel() {
+        return answerModel;
     }
 
     public QuestionVoteModel getVotes() {
         return votes;
-    }
-
-    public String getPostMessage() {
-        return postMessage;
-    }
-
-    public void setPostMessage(String postMessage) {
-        this.postMessage = postMessage;
-    }
-
-    public String getDateTime() {
-        return dateTime;
-    }
-
-    public void setDateTime(String dateTime) {
-        this.dateTime = dateTime;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
     }
 }
