+++ src/main/java/backend/model/vote/AnswerVoteModel.java
@@ -1,10 +1,9 @@
 package backend.model.vote;
 
 import backend.model.qa.AnswerModel;
+import backend.model.vote.common.VoteModel;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
 import javax.persistence.OneToOne;
 
 /**
@@ -12,20 +11,13 @@ import javax.persistence.OneToOne;
  */
 
 @Entity
-public class AnswerVoteModel  {
+public class AnswerVoteModel extends VoteModel{
 
-    @Id
-    @GeneratedValue
-    private long id;
     @OneToOne
     AnswerModel answer;
-    private int upVotes;
-    private int downVotes;
 
     public AnswerVoteModel(AnswerModel answer) {
         this.answer = answer;
-        upVotes = 1;
-        downVotes = 0;
     }
 
     public AnswerVoteModel(){
@@ -35,35 +27,4 @@ public class AnswerVoteModel  {
         return answer;
     }
 
-    public long getId() {
-        return id;
-    }
-
-    public int getUpVotes() {
-        return upVotes;
-    }
-
-    public int getDownVotes() {
-        return downVotes;
-    }
-
-    public int getTotalVotes() {
-        return upVotes + downVotes;
-    }
-
-    public void decrementUpVotes() {
-        upVotes--;
-    }
-
-    public void incrementUpVotes() {
-        upVotes++;
-    }
-
-    public void decrementDownVotes() {
-        downVotes--;
-    }
-
-    public void incrementDownVotes() {
-        downVotes++;
-    }
 }
